backend/requirements.txt
# Core FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Image processing and OCR
opencv-python==4.8.1.78
pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
numpy==1.25.2

# Machine Learning and NLP
torch==2.1.1
torchvision==0.16.1
transformers==4.35.2
scikit-learn==1.3.2
spacy==3.7.2
fuzzywuzzy==0.18.0
python-Levenshtein==0.23.0

# HTTP clients and external APIs
httpx==0.25.2
aiohttp==3.9.1
requests==2.31.0

# Data processing
pandas==2.1.3
pydantic==2.5.0
pydantic-settings==2.1.0

# Monitoring and logging
structlog==23.2.0
sentry-sdk[fastapi]==1.38.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development tools
black==23.11.0
isort==5.12.0
mypy==1.7.1
pre-commit==3.6.0

# Production
gunicorn==21.2.0
redis==5.0.1
celery==5.3.4